Refactor: extrac color function
